// GC-Mobile-Web-Application/UnitedAdminPortal/src/app/models/user.model.ts
export class User {
    UserID: string;
    FirstName: string;
    LastName: string;
    username: string;
    password: string;
    StreetAddress: string;
    City: string;
    Country: string;
    Zipcode: string;
    ContactNumber: string;
    Email: string;
    profilePictureLink: string;
    Preference1: string;
    Preference2: string;
    Preference3: string;
}